+++ R/import-googlesheets.R
@@ -75,6 +75,7 @@ import_googlesheets_ui <- function(id, title = TRUE) {
 #'  \code{"button"} (when user click on button) or
 #'  \code{"change"} (each time user select a dataset in the list).
 #' @param return_class Class of returned data: \code{data.frame}, \code{data.table} or \code{tbl_df} (tibble).
+#' @param reset A `reactive` function that when triggered resets the data.
 #'
 #' @export
 #'
@@ -87,7 +88,8 @@ import_googlesheets_ui <- function(id, title = TRUE) {
 import_googlesheets_server <- function(id,
                                        btn_show_data = TRUE,
                                        trigger_return = c("button", "change"),
-                                       return_class = c("data.frame", "data.table", "tbl_df")) {
+                                       return_class = c("data.frame", "data.table", "tbl_df"),
+                                       reset = reactive(NULL)) {
 
   trigger_return <- match.arg(trigger_return)
 
@@ -96,6 +98,12 @@ import_googlesheets_server <- function(id,
     ns <- session$ns
     imported_rv <- reactiveValues(data = NULL, name = NULL)
     temporary_rv <- reactiveValues(data = NULL, name = NULL, status = NULL)
+
+    observeEvent(reset(), {
+      temporary_rv$data <- NULL
+      temporary_rv$name <- NULL
+      temporary_rv$status <- NULL
+    })
 
     output$container_confirm_btn <- renderUI({
       if (identical(trigger_return, "button")) {