+++ NEWS.md
@@ -1,3 +1,9 @@
+# datamods 1.1.5
+
+* `import_*_server()` added reset argument to clear the data.
+* `import_copypaste_server()` also return a `reative` function "name" lile the others.
+
+
 # datamods 1.1.4
 
 * `filter_data_server`: convert data to `data.frame` (fix [esquisse #149](https://github.com/dreamRs/esquisse/issues/149)).